Add more specific string literal matching
Some of the patterns are picking up code comments, and unrelated
english words - for example javadoc comment for DESCRIPTION of an sql
schema.

Ensure that relevant digests literal strings are looked up as part of
a string, ie. "MD5" 'MD5' 'MD5WithRsa' and so on. This somewhat
reduces false positive matches, but possibly increases false
negatives.

Possibly need to add a range of patterns, i.e. apache commons can
access digest by string, but also with a specific method name or
constant. Such that maybe matching should be added for getSha1Digest()
ahd SHA_1, however not sure how useful that is as all of apache
commons codecs is not a certified implementation.